+++ Source/DreamShaderEditor/Private/DreamShaderEditorBridge.h
@@ -55,6 +55,7 @@ namespace UE::DreamShader::Editor::Private
 		void PopulateMaterialFunctionDreamShaderMenu(UToolMenu* InMenu, TWeakObjectPtr<UMaterialFunction> MaterialFunction);
 		void RequestRecompileAll();
 		void RequestCleanGeneratedShaders();
+		void OpenDreamShaderWorkspace();
 		void CopyVirtualFunctionDefinition(TWeakObjectPtr<UMaterialFunction> MaterialFunction);
 		void CreateVirtualFunctionDefinitionFile(TWeakObjectPtr<UMaterialFunction> MaterialFunction);
 		void CopyVirtualFunctionCall(TWeakObjectPtr<UMaterialFunction> MaterialFunction);